Fix tests. Refactor exceptions
