+++ __tests__/services/users.service.spec.ts
@@ -5,8 +5,9 @@ import { factory } from 'typeorm-seeding';
 import { User } from '@entities/user.entity';
 import { getRepository, Repository } from 'typeorm';
 import { mockUpdateResult } from '../utils/mocks';
-import { HashInvalidError } from '@exception/users/hashinvalid.error';
-import { HashExpiredError } from '@exception/users/hashexpired.error';
+import { HashInvalidError } from '@exception/users/hash-invalid.error';
+import { HashExpiredError } from '@exception/users/hash-expired.error';
+import * as faker from 'faker';
 
 let usersService: UsersService;
 let user: User;
@@ -15,11 +16,14 @@ let userRepository: Repository<User>;
 describe('UsersService', () => {
   beforeAll(async () => {
     usersService = Container.get(UsersService);
+    userRepository = getRepository<User>(User);
   });
 
   it('all dependencies should be defined', () => {
     expect(usersService).toBeDefined();
+    expect(userRepository).toBeDefined();
   });
+
   describe('comparePassword', () => {
     let userPassword: string;
 
@@ -42,33 +46,29 @@ describe('UsersService', () => {
       expect(result).toBeFalsy();
     });
   });
+
   describe('verifyUser', () => {
     beforeEach(async () => {
       user = await factory(User)().create();
     });
 
-    beforeAll( () => {
-      usersService = Container.get(UsersService);
-      userRepository = getRepository<User>(User);
-    });
-
     it('should throw error if the hash is invalid', async () => {
-      user.verifyHash = '11111111-2222-3333-4444-555555555555';
+      user.verifyHash = faker.datatype.uuid();
 
       jest.spyOn(usersService, 'showUserByHash')
         .mockRejectedValueOnce(new HashInvalidError);
 
-      expect(usersService.verifyUser(user.verifyHash))
+      await expect(usersService.verifyUser(user.verifyHash))
         .rejects.toThrowError(HashInvalidError);
     });
 
     it('should throw error if the hash is expired', async () => {
-      user.hashExpiresAt = new Date();
+      user.hashExpiresAt = faker.date.past();
 
       jest.spyOn(usersService, 'showUserByHash')
         .mockResolvedValueOnce(user);
 
-      expect(usersService.verifyUser(user.verifyHash))
+      await expect(usersService.verifyUser(user.verifyHash))
         .rejects.toThrowError(HashExpiredError);
     });
 
@@ -79,9 +79,9 @@ describe('UsersService', () => {
         .mockResolvedValueOnce(mockUpdateResult);
 
       const userVerified = await usersService.verifyUser(user.verifyHash);
-      expect(userVerified).toHaveProperty('verified', true);
-      expect(userVerified).toHaveProperty('verifyHash', null);
-      expect(userVerified).toHaveProperty('hashExpiresAt', null);
+      expect(userVerified.verified).toBeTruthy();
+      expect(userVerified.verifyHash).toBeNull();
+      expect(userVerified.hashExpiresAt).toBeNull();
     });
   });
 });
